order of widgets to float to allow tweaking

+++ beegisplugins/eu.hydrologis.jgrass.formeditor/src/eu/hydrologis/jgrass/featureeditor/xml/annotated/AForm.java
@@ -48,7 +48,7 @@ public class AForm extends FormElement {
     }
 
     @Override
-    public int getOrder() {
+    public float getOrder() {
         return -1;
     }
